Add eslint config airbnb
