+++ .eslintignore
@@ -1,0 +1,2 @@
+src/registerServiceWorker.js
+src/App.test.js
