+++ .eslintrc.json
@@ -1,0 +1,6 @@
+{
+    "extends": "airbnb",
+    "rules": {
+        "react/prop-types": 0
+    }
+}
