+++ src/Complexes/Show/Offer.jsx
@@ -1,6 +1,6 @@
-import React from "react";
-import { Col } from "react-flexbox-grid";
-import styled from "styled-components";
+import React from 'react';
+import { Col } from 'react-flexbox-grid';
+import styled from 'styled-components';
 
 const Offer = styled.div`
   padding-top: 1.5rem;
@@ -58,23 +58,21 @@ const Button = styled.button`
   cursor: pointer;
 `;
 
-export default props => {
-  return (
-    <Col lg={4}>
-      <Offer>
-        <Title>{props.title}</Title>
-        <Info>
-          <Label>Площадь</Label>
-          <Value>от {props.area.min} до {props.area.max} м²</Value>
-        </Info>
-        <Info>
-          <Label>Стоимость</Label>
-          <Value>от {props.price.min} до {props.price.max} млн.руб</Value>
-        </Info>
-        <Button>
+export default props => (
+  <Col lg={4}>
+    <Offer>
+      <Title>{props.title}</Title>
+      <Info>
+        <Label>Площадь</Label>
+        <Value>от {props.area.min} до {props.area.max} м²</Value>
+      </Info>
+      <Info>
+        <Label>Стоимость</Label>
+        <Value>от {props.price.min} до {props.price.max} млн.руб</Value>
+      </Info>
+      <Button>
           Посмотреть предложения
         </Button>
-      </Offer>
-    </Col>
+    </Offer>
+  </Col>
   );
-};
